# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/whiz/tss/sspark/s_spark_android/presentation/menu/MenuStudentFragment.kt
#	library/src/main/java/whiz/sspark/library/data/enum/MenuCode.kt
#	library/src/main/res/values-th/strings.xml
#	library/src/main/res/values/strings.xml

+++ library/src/main/java/whiz/sspark/library/utility/DateTimeUtility.kt
@@ -31,7 +31,7 @@ fun getLatestUpdatedDateTime(context: Context, date: Date): String {
     return context.resources.getString(R.string.screen_header_subtitle, dateTimeFormat)
 }
 
-fun convertToLocalizeYear(year: Long): String {
+fun convertToLocalizeYear(year: Int): String {
     return if (isThaiLanguage()) {
         (year + 543).toString()
     } else {
